Refactor landing card click event handling and add console log statement
Adjust the landing card click event handling in LandingCard.jsx to include a console log statement when the card is clicked. This change allows for better debugging and understanding of user interactions.

Also, refactor hexagon size and spacing in Background.jsx

In Background.jsx, update the size and spacing of the hexagons to improve the visual appearance. The width and height of the hexagons are adjusted, as well as the vertical and horizontal spacing between them. This change enhances the overall design and layout of the background pattern.

Refactor code formatting and add missing newlines in LandingCard.js, Projects.css, and Projects.js

Improve code readability and maintainability by refactoring the code formatting in LandingCard.js, Projects.css, and Projects.js. Additionally, add missing newlines to ensure consistent code structure.